Allow groups to be fetched by ID or Name
Allow groups to be fetched and deleted by either ID or name.
The user can hit the endpoint /groups/:ident with either an
ID of a group or a name, first the route will look for groups
matching the ID and if the ident doesn't look like an ID
it will check for names.